Add DISPLAY support
This is a bit iffy, as Xlib typically uses abstract unix domain
sockets to connect to X, but that is not namespaced. We instead set
DISPLAY=99 and point /tmp/.X11-unix/X99 to the right X socket. Any
Xserver listening to global abstract unix domain sockets are still
accessible to the app though...